style: Foldering 및 코드 형식 수정
