--- tmdb project/tmdb project/Controller/SearchController.swift
@@ -1,12 +1,0 @@
-//
-//  SearchController.swift
-//  tmdb project
-//
-//  Created by 박태현 on 2022/02/27.
-//
-
-import UIKit
-
-class SearchController: UISearchController {
-    
-}
