+++ tmdb project/tmdb project/Extensions/UILabel+CommonForm.swift
@@ -1,0 +1,15 @@
+//
+//  UILabel+CommonForm.swift
+//  tmdb project
+//
+//  Created by 박태현 on 2022/02/28.
+//
+
+import UIKit
+
+extension UILabel {
+    func settingCommonForm() {
+        self.textColor = .white
+        self.font = UIFont.systemFont(ofSize: 14)
+    }
+}
