+++ tmdb project/tmdb project/Model/Dates.swift
@@ -1,0 +1,13 @@
+//
+//  Dates.swift
+//  tmdb project
+//
+//  Created by 박태현 on 2022/03/01.
+//
+
+import Foundation
+
+struct Dates: Decodable {
+    let maximum: String
+    let minimum: String
+}
