+++ tmdb project/tmdb project/View/SearchCollectionViewCell.swift
@@ -20,15 +20,13 @@ class SearchCollectionViewCell: UICollectionViewCell {
 
     private lazy var titleLabel: UILabel = {
         let label = UILabel()
-        label.textColor = .white
-        label.font = UIFont.systemFont(ofSize: 14)
+        label.settingCommonForm()
         return label
     }()
 
     private lazy var releaseDateLabel: UILabel = {
         let label = UILabel()
-        label.textColor = .white
-        label.font = UIFont.systemFont(ofSize: 14)
+        label.settingCommonForm()
         return label
     }()
 
@@ -40,8 +38,7 @@ class SearchCollectionViewCell: UICollectionViewCell {
 
     private lazy var popularityLabel: UILabel = {
         let label = UILabel()
-        label.textColor = .white
-        label.font = UIFont.systemFont(ofSize: 14)
+        label.settingCommonForm()
         return label
     }()
